refactor: config: move from cmd/acp to pkg/config
